Refactor logic for normal and amend status.
Previously, I was using the `NodeGit.Repository.getStatus` API to get "normal"
status, and the direct `NodeGit.Diff` routines for getting amend status.  It
turns out that the `getStatus` API isn't really sufficient even for the normal
case: you can't, for example, detect that a workdir change cancels out a staged
change.
