+++ node/lib/util/commit.js
@@ -46,7 +46,6 @@ const RepoStatus          = require("./repo_status");
 const PrintStatusUtil     = require("./print_status_util");
 const SubmoduleFetcher    = require("./submodule_fetcher");
 const SubmoduleUtil       = require("./submodule_util");
-const SubmoduleConfigUtil = require("./submodule_config_util");
 
 /**
  * Commit changes in the specified `repo`.  If the specified `doAll` is true,
@@ -288,96 +287,6 @@ exports.commit = co.wrap(function *(metaRepo, all, metaStatus, message) {
     return null;
 });
 
-function readDiff(diff) {
-    const result = {};
-    const FILESTATUS = RepoStatus.FILESTATUS;
-    const numDeltas = diff.numDeltas();
-    for (let i = 0;  i < numDeltas; ++i) {
-        const delta = diff.getDelta(i);
-        const diffStatus = delta.status();
-        const fileStatus = DiffUtil.convertDeltaFlag(diffStatus);
-        const file = FILESTATUS.REMOVED === fileStatus ?
-                     delta.oldFile() :
-                     delta.newFile();
-        const path = file.path();
-
-        // Skip the .gitmodules file and all submodule changes; they're handled
-        // separately.
-
-        if (SubmoduleConfigUtil.modulesFileName !== path &&
-            NodeGit.TreeEntry.FILEMODE.COMMIT !== file.mode()) {
-            result[path] = fileStatus;
-        }
-    }
-    return result;
-}
-
-/**
- * Return the `staged` and `workdir` files for the specified `repo` having the
- * current specified `status`, indicating the changes to be made against the
- * base commit.  If the specified `all` is true, compare the workdir against
- * the base tree; this may be different than the diff against the index
- * indicated in `status`.  Create the diff agains tthe specified `parentTree`,
- * if not null.
- *
- * @param {NodeGit.Repository} repo
- * @param {RepoStatus}         status
- * @param {NodeGit.Tree|null}  parentTree
- * @param {Boolena}            all
- * @return {Object}
- * @return {Object} return.staged path to FILESTATUS
- * @return {Object} return.workdir path to FILESTATUS
- */
-exports.getAmendStatusForRepo = co.wrap(function *(repo,
-                                                   status,
-                                                   parentTree,
-                                                   all) {
-    // When we're just doing staged files (i.e., not 'all'), the problem is
-    // pretty straightforward: we diff the index to the tree and return that as
-    // 'staged'; 'workdir' is always empty.
-
-    const options = { ignoreSubmodules: 1, };
-    const indexDiff =
-               yield NodeGit.Diff.treeToIndex(repo, parentTree, null, options);
-    const indexFiles = readDiff(indexDiff);
-    const workdirStatus = status.workdir;
-    if (!all) {
-        return {
-            staged: indexFiles,
-            workdir: workdirStatus,
-        };
-    }
-    const workdirDiff =
-                   yield NodeGit.Diff.treeToWorkdir(repo, parentTree, options);
-    const workdirFiles = readDiff(workdirDiff);
-    const staged = {};
-    const workdir = {};
-
-    // When doing all, the algorithm is to use the index status only for files
-    // that have no workdir change -- unless the workdir change is an addition.
-
-    const ADDED = RepoStatus.FILESTATUS.ADDED;
-
-    Object.keys(indexFiles).forEach(path => {
-        const workdir = workdirStatus[path];
-        if (undefined === workdir || ADDED === workdir) {
-            staged[path] = indexFiles[path];
-        }
-    });
-
-    Object.keys(workdirFiles).forEach(path => {
-        const status = workdirStatus[path];
-        if (undefined !== status) {
-            workdir[path] = workdirFiles[path];
-        }
-    });
-
-    return {
-        staged: staged,
-        workdir: workdir,
-    };
-});
-
 /**
  * Return true if the specified `x` and `y` commits appear to have been
  * generated by the same `git meta commit` invocation; namely that the
@@ -493,6 +402,12 @@ const getHeadParentTree = co.wrap(function *(repo) {
     return yield NodeGit.Tree.lookup(repo, treeId);
  });
 
+const getAmendStatusForRepo = co.wrap(function *(repo, all) {
+    const tree = yield getHeadParentTree(repo);
+    return yield DiffUtil.getRepoStatus(repo, tree, [], all, true);
+});
+
+
 /**
  * Return a list of submodules to be amended and a new `RepoStatus` object
  * reflecting all changes to be part of an amend commit (including the
@@ -529,9 +444,7 @@ exports.getAmendChanges = co.wrap(function *(repo,
     let metaStaged = {};
     let metaWorkdir = {};
     if (includeMeta) {
-        const tree = yield getHeadParentTree(repo);
-        const changes =
-            yield exports.getAmendStatusForRepo(repo, status, tree, all);
+        const changes = yield getAmendStatusForRepo(repo, all);
         metaStaged = changes.staged;
         metaWorkdir = changes.workdir;
     }
@@ -560,11 +473,7 @@ exports.getAmendChanges = co.wrap(function *(repo,
         }
 
         const subRepo = yield SubmoduleUtil.getRepo(repo, subName);
-        const tree = yield getHeadParentTree(subRepo);
-        const changes = yield exports.getAmendStatusForRepo(subRepo,
-                                                            newSub.repoStatus,
-                                                            tree,
-                                                            all);
+        const changes = yield getAmendStatusForRepo(subRepo, all);
         subsToAmend.push(subName);
 
         // Now, we need to update the repo status of the submodule to reflect
