+++ node/test/util/commit.js
@@ -336,150 +336,6 @@ x=E:Cx-1 s=Sa:s;I s=~;Os Cs q=r!*=master!Bmaster=s;Bmaster=x`,
         });
     });
 
-    describe("getAmendStatusForRepo", function () {
-        const FILESTATUS = RepoStatus.FILESTATUS;
-        const cases = {
-            "first commit": {
-                input: "x=S",
-                staged: { "README.md": FILESTATUS.ADDED},
-            },
-            "added": {
-                input: "x=S:W foo=bar",
-                staged: { "README.md": FILESTATUS.ADDED},
-                workdir: { foo: FILESTATUS.ADDED },
-            },
-            "modified, not all": {
-                input: "x=S:C2-1;W README.md=3;Bmaster=2",
-                staged: { "2": FILESTATUS.ADDED },
-                workdir: { "README.md": FILESTATUS.MODIFIED },
-            },
-            "staged modified": {
-                input: "x=S:C2-1;W README.md=3;Bmaster=2",
-                staged: { "2": FILESTATUS.ADDED },
-                workdir: { "README.md": FILESTATUS.MODIFIED },
-                all: true,
-            },
-            "changed in index": {
-                input: "x=S:C2-1;I README.md=3;Bmaster=2",
-                staged: {
-                    "2": FILESTATUS.ADDED,
-                    "README.md": FILESTATUS.MODIFIED,
-                },
-            },
-            "added in index": {
-                input: "x=S:C2-1;I foo=3;Bmaster=2",
-                staged: {
-                    "2": FILESTATUS.ADDED,
-                    "foo": FILESTATUS.ADDED,
-                },
-            },
-            "removed in index": {
-                input: "x=S:C2-1;I README.md;Bmaster=2",
-                staged: {
-                    "2": FILESTATUS.ADDED,
-                    "README.md": FILESTATUS.REMOVED,
-                },
-            },
-            "removed in index with all": {
-                input: "x=S:C2-1;I README.md;Bmaster=2",
-                staged: {
-                    "2": FILESTATUS.ADDED,
-                    "README.md": FILESTATUS.REMOVED,
-                },
-                all: true,
-            },
-            "removed in workdir": {
-                input: "x=S:C2-1;W README.md;Bmaster=2",
-                staged: {
-                    "2": FILESTATUS.ADDED,
-                },
-                workdir: {
-                    "README.md": FILESTATUS.REMOVED,
-                },
-            },
-            "removed in workdir with all": {
-                input: "x=S:C2-1;W README.md;Bmaster=2",
-                staged: {
-                    "2": FILESTATUS.ADDED,
-                },
-                workdir: {
-                    "README.md": FILESTATUS.REMOVED,
-                },
-                all: true,
-            },
-            "ignore submodule add": {
-                input: `
-a=B:Ca-1;Bmaster=a|
-x=S:C2-1 s=Sa:1;Bmaster=2`,
-            },
-            "ignore submodule change": {
-                input: `
-a=B:Ca-1;Bmaster=a|
-x=S:C2-1 s=Sa:1;C3-2 s=Sa:a;Bmaster=3`,
-            },
-            "ignore submodule deletion": {
-                input: `
-a=B:Ca-1;Bmaster=a|
-x=S:C2-1 s=Sa:1;C3-2 s;Bmaster=3`,
-            },
-            "unmodified": {
-                input: `
-x=S:C2-1 README.md=3;W README.md=hello world;Bmaster=2`,
-                all: true,
-            },
-            "changed in index, rm'd in workdir": {
-                input: "x=S:C2-1;I README.md=3;W README.md;Bmaster=2",
-                staged: {
-                    "2": FILESTATUS.ADDED,
-                    "README.md": FILESTATUS.MODIFIED
-                },
-                workdir: {
-                    "README.md": FILESTATUS.REMOVED,
-                },
-            },
-            "changed in index, rm'd in workdir, all": {
-                input: "x=S:C2-1;I README.md=3;W README.md;Bmaster=2",
-                staged: {
-                    "2": FILESTATUS.ADDED,
-                },
-                workdir: {
-                    "README.md": FILESTATUS.REMOVED,
-                },
-                all: true,
-            },
-        };
-        Object.keys(cases).forEach(caseName => {
-            const c = cases[caseName];
-            it(caseName, co.wrap(function *() {
-                const written =
-                               yield RepoASTTestUtil.createMultiRepos(c.input);
-                const repos = written.repos;
-                const repo = repos.x;
-                const status = yield StatusUtil.getRepoStatus(repo, {
-                    showMetaChanges: true,
-                    includeClosedSubmodules: true,
-                });
-                const head = yield repo.getHeadCommit();
-                let tree = null;
-                if (0 !== head.parentcount()) {
-                    const parent = yield head.parent(0);
-                    const treeId = parent.treeId();
-                    tree = yield NodeGit.Tree.lookup(repo, treeId);
-                }
-                const result = yield Commit.getAmendStatusForRepo(
-                                                               repo,
-                                                               status,
-                                                               tree,
-                                                               c.all || false);
-                const expected = {
-                    staged: c.staged || {},
-                    workdir: c.workdir || {},
-                };
-                assert.deepEqual(result, expected);
-            }));
-        });
-    });
-
     describe("sameCommitInstance", function () {
         const cases = {
             "same weird": {
