+++ node/test/util/status_util.js
@@ -193,180 +193,6 @@ describe("StatusUtil", function () {
         });
     });
 
-    describe("getChanges", function () {
-        const cases = {
-            "trivial": {
-                state: "x=S",
-            },
-            "index - modified": {
-                state: "x=S:I README.md=hhh",
-                staged: { "README.md": FILESTATUS.MODIFIED },
-            },
-            "index - modified deep": {
-                state: "x=S:C2 x/y/z=a;I x/y/z=b;H=2",
-                staged: { "x/y/z": FILESTATUS.MODIFIED },
-            },
-            "index - added": {
-                state: "x=S:I x=y",
-                staged: { x: FILESTATUS.ADDED },
-            },
-            "index - added deep": {
-                state: "x=S:I x/y=y",
-                staged: { "x/y": FILESTATUS.ADDED },
-            },
-            "index - removed": {
-                state: "x=S:I README.md",
-                staged: { "README.md": FILESTATUS.REMOVED},
-            },
-            "index - removed deep": {
-                state: "x=S:C2 x/y/z=a;I x/y/z;H=2",
-                staged: { "x/y/z": FILESTATUS.REMOVED},
-            },
-            "workdir - modified": {
-                state: "x=S:W README.md=hhh",
-                workdir: { "README.md": FILESTATUS.MODIFIED },
-            },
-            "workdir - modified deep": {
-                state: "x=S:C2 x/y/z=a;W x/y/z=b;H=2",
-                workdir: { "x/y/z": FILESTATUS.MODIFIED },
-            },
-            "workdir - added": {
-                state: "x=S:W x=y",
-                workdir: { x: FILESTATUS.ADDED },
-            },
-            "workdir - added deep": {
-                state: "x=S:W x/y=y",
-                workdir: { "x/": FILESTATUS.ADDED },
-            },
-            "workdir - added deep all untracked": {
-                state: "x=S:W x/y=y",
-                allUntracked: true,
-                workdir: { "x/y": FILESTATUS.ADDED },
-            },
-            "workdir - removed": {
-                state: "x=S:W README.md",
-                workdir: { "README.md": FILESTATUS.REMOVED},
-            },
-            "workdir - removed deep": {
-                state: "x=S:C2 x/y/z=a;W x/y/z;H=2",
-                workdir: { "x/y/z": FILESTATUS.REMOVED},
-            },
-            "modified workdir and index": {
-                state: "x=S:I README.md=aaa;W README.md=bbb",
-                staged: { "README.md": FILESTATUS.MODIFIED },
-                workdir: { "README.md": FILESTATUS.MODIFIED },
-            },
-            "index path restriction": {
-                state: "x=S:I README.md=aaa,foo=a",
-                paths: [ "foo" ],
-                staged: { foo: FILESTATUS.ADDED },
-            },
-            "index dir path": {
-                state: "x=S:I x/y/z=foo,x/r/z=bar,README.md",
-                paths: [ "x" ],
-                staged: {
-                    "x/y/z": FILESTATUS.ADDED,
-                    "x/r/z": FILESTATUS.ADDED
-                },
-            },
-            "index dir paths": {
-                state: "x=S:I x/y/z=foo,x/r/z=bar,README.md",
-                paths: [ "x/y", "x/r" ],
-                staged: {
-                    "x/y/z": FILESTATUS.ADDED,
-                    "x/r/z": FILESTATUS.ADDED
-                },
-            },
-            "index all paths": {
-                state: "x=S:I x/y/z=foo,x/r/z=bar,README.md",
-                paths: [ "x/y/z", "x/r/z", "README.md" ],
-                staged: {
-                    "x/y/z": FILESTATUS.ADDED,
-                    "x/r/z": FILESTATUS.ADDED,
-                    "README.md": FILESTATUS.REMOVED,
-                },
-            },
-            "workdir path restriction": {
-                state: "x=S:W README.md=aaa,foo=a",
-                paths: [ "foo" ],
-                workdir: { foo: FILESTATUS.ADDED },
-            },
-            "workdir dir path": {
-                state: "x=S:W x/y/z=foo,x/r/z=bar,README.md",
-                paths: [ "x" ],
-                allUntracked: true,
-                workdir: {
-                    "x/y/z": FILESTATUS.ADDED,
-                    "x/r/z": FILESTATUS.ADDED
-                },
-            },
-            "workdir dir paths": {
-                state: "x=S:W x/y/z=foo,x/r/z=bar,README.md",
-                paths: [ "x/y", "x/r" ],
-                allUntracked: true,
-                workdir: {
-                    "x/y/z": FILESTATUS.ADDED,
-                    "x/r/z": FILESTATUS.ADDED
-                },
-            },
-            "workdir all paths": {
-                state: "x=S:W x/y/z=foo,x/r/z=bar,README.md",
-                paths: [ "x/y/z", "x/r/z", "README.md" ],
-                allUntracked: true,
-                workdir: {
-                    "x/y/z": FILESTATUS.ADDED,
-                    "x/r/z": FILESTATUS.ADDED,
-                    "README.md": FILESTATUS.REMOVED,
-                },
-            },
-            "many changes": {
-                state: `
-x=S:C2 a/b=c,a/c=d,t=u;H=2;I a/b,a/q=r,f=x;W a/b=q,a/c=f,a/y=g,f`,
-                allUntracked: true,
-                workdir: {
-                    "a/b": FILESTATUS.ADDED,
-                    "a/c": FILESTATUS.MODIFIED,
-                    "a/y": FILESTATUS.ADDED,
-                    "f": FILESTATUS.REMOVED,
-                },
-                staged: {
-                    "a/b": FILESTATUS.REMOVED,
-                    "a/q": FILESTATUS.ADDED,
-                    "f": FILESTATUS.ADDED,
-                },
-            },
-            "many changes with path": {
-                state: `
-x=S:C2 a/b=c,a/c=d,t=u;H=2;I a/b,a/q=r,f=x;W a/b=q,a/c=f,a/y=g,f`,
-                allUntracked: true,
-                paths: ["f"],
-                workdir: {
-                    "f": FILESTATUS.REMOVED,
-                },
-                staged: {
-                    "f": FILESTATUS.ADDED,
-                },
-            },
-        };
-        Object.keys(cases).forEach(caseName => {
-            const c = cases[caseName];
-            it(caseName, co.wrap(function *() {
-                const w = yield RepoASTTestUtil.createMultiRepos(c.state);
-                const repo = w.repos.x;
-                const paths = c.paths || [];
-                const allUntracked = c.allUntracked || false;
-                const result = yield StatusUtil.getChanges(repo,
-                                                           paths,
-                                                           allUntracked);
-                const expected = {
-                    staged: c.staged || {},
-                    workdir: c.workdir || {},
-                };
-                assert.deepEqual(result, expected);
-            }));
-        });
-    });
-
     describe("getSubmoduleStatus", function () {
         // We will use `x` for the repo name and `s` for the submodule name.
